Move date parsing and validation into separate module
