+++ lib/masamune/actions.rb
@@ -7,6 +7,7 @@ module Masamune
     require 'masamune/actions/postgres'
     require 'masamune/actions/postgres_admin'
     require 'masamune/actions/filesystem'
+    require 'masamune/actions/date_parse'
     require 'masamune/actions/data_flow'
     require 'masamune/actions/elastic_mapreduce'
     require 'masamune/actions/execute'
